feat: solve streams, modules, hash
