+++ src/hash/calcHash.js
@@ -1,5 +1,12 @@
+import { createReadStream } from "node:fs";
+import { stdout } from "node:process";
+const { createHash } = await import("node:crypto");
+
 const calculateHash = async () => {
-    // Write your code here 
+    const hash = createHash("sha256");
+
+    const input = createReadStream(import.meta.dirname + "/files/fileToCalculateHashFor.txt");
+    input.pipe(hash).setEncoding("hex").pipe(stdout);
 };
 
 await calculateHash();
