+++ src/streams/files/fileToWrite.txt
@@ -1,0 +1,1 @@
+holamola
