+++ src/streams/read.js
@@ -1,5 +1,9 @@
+import { createReadStream } from "node:fs";
+import { stdout } from "node:process";
+
 const read = async () => {
-    // Write your code here 
+    const input = createReadStream(import.meta.dirname + "/files/fileToRead.txt");
+    input.pipe(stdout).setEncoding("hex");
 };
 
 await read();
