+++ src/streams/transform.js
@@ -1,5 +1,14 @@
+import { Transform } from "node:stream";
+import { stdout, stdin } from "node:process";
 const transform = async () => {
-    // Write your code here 
+    const reverse = new Transform({
+        transform: function (data, enc, cb) {
+            this.push(data.toString().trim().split("").reverse().join("") + "\n");
+            cb();
+        },
+    });
+
+    stdin.pipe(reverse).pipe(stdout);
 };
 
 await transform();
